pyproject.toml
[project]
name = "openai-api-polling"
version = "0.1.2"
description = "Polling OpenAI API without rate limit issues."
readme = "README.md"
authors = [
    { name = "SongTan", email = "sepinetam@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.75.0",
    "google-genai>=1.55.0",
    "openai>=2.0.0",
    "tiktoken>=0.10.0",
]

[project.scripts]
openai-api-polling = "openai_api_polling:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://github.com/sepinetam/openai-api-polling"
Source = "https://github.com/sepinetam/openai-api-polling"
